MemSharingHypervisor: Use usize for base_ipa
base_ipa is a physical address so use the more appropriate type.

Test: m pvmfw_bin service_vm_bin vmbase_example_bios_bin
Flag: EXEMPT PURE_REFACTOR

+++ libs/libhypervisor_backends/src/hypervisor/kvm_x86.rs
@@ -139,7 +139,7 @@ macro_rules! cpuid {
 }
 
 impl MemSharingHypervisor for ProtectedKvmHypervisor {
-    fn share(&self, base_ipa: u64) -> Result<()> {
+    fn share(&self, base_ipa: usize) -> Result<()> {
         let ret: u32 = vmcall!(PKVM_GHC_SHARE_MEM, base_ipa, SIZE_4KB);
 
         if ret != 0 {
@@ -149,7 +149,7 @@ impl MemSharingHypervisor for ProtectedKvmHypervisor {
         Ok(())
     }
 
-    fn unshare(&self, base_ipa: u64) -> Result<()> {
+    fn unshare(&self, base_ipa: usize) -> Result<()> {
         let ret: u32 = vmcall!(PKVM_GHC_UNSHARE_MEM, base_ipa, SIZE_4KB);
         if ret != 0 {
             return Err(Error::KvmError(KvmError::from(ret as i32), PKVM_GHC_UNSHARE_MEM));
